python-objstore/requirements.txt
minio==7.1.0
python-dotenv==0.19.0
boto3==1.18.28
